lsm-testing/requirements.dev.txt
inmanta-dev-dependencies[module]==2.113.0
pytest-inmanta==2.9.0
pytest-inmanta-lsm==3.3.0
types-PyYAML==6.0.12.20240311
